pyproject.toml
[project]
name = "agent1001"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "openai>=1.99.5",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
]

[tool.uv.pip]
index-url = "https://pypi.tuna.tsinghua.edu.cn/simple"
